feat: refactor wasm bindings structure
