+++ x/amm/client/wasm/querier/query_swap_estimation.go
@@ -3,9 +3,9 @@ package wasm
 import (
 	wasmvmtypes "github.com/CosmWasm/wasmvm/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	wasmbindingstypes "github.com/elys-network/elys/wasmbindings/types"
+	"github.com/elys-network/elys/x/amm/client/wasm/types"
 )
 
-func (oq *Querier) querySwapEstimation(ctx sdk.Context, query *wasmbindingstypes.QuerySwapEstimationRequest) ([]byte, error) {
+func (oq *Querier) querySwapEstimation(ctx sdk.Context, query *types.QuerySwapEstimationRequest) ([]byte, error) {
 	return nil, wasmvmtypes.UnsupportedRequest{Kind: "QuerySwapEstimation, not implemented yet"}
 }
