+++ x/margin/client/wasm/messenger/msg_close.go
@@ -6,12 +6,12 @@ import (
 	errorsmod "cosmossdk.io/errors"
 	wasmvmtypes "github.com/CosmWasm/wasmvm/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	wasmbindingstypes "github.com/elys-network/elys/wasmbindings/types"
+	"github.com/elys-network/elys/x/margin/client/wasm/types"
 	marginkeeper "github.com/elys-network/elys/x/margin/keeper"
 	margintypes "github.com/elys-network/elys/x/margin/types"
 )
 
-func (m *Messenger) msgClose(ctx sdk.Context, contractAddr sdk.AccAddress, msgClose *wasmbindingstypes.MsgClose) ([]sdk.Event, [][]byte, error) {
+func (m *Messenger) msgClose(ctx sdk.Context, contractAddr sdk.AccAddress, msgClose *types.MsgClose) ([]sdk.Event, [][]byte, error) {
 	res, err := PerformMsgClose(m.keeper, ctx, contractAddr, msgClose)
 	if err != nil {
 		return nil, nil, errorsmod.Wrap(err, "perform close")
@@ -27,7 +27,7 @@ func (m *Messenger) msgClose(ctx sdk.Context, contractAddr sdk.AccAddress, msgCl
 	return nil, resp, nil
 }
 
-func PerformMsgClose(f *marginkeeper.Keeper, ctx sdk.Context, contractAddr sdk.AccAddress, msgClose *wasmbindingstypes.MsgClose) (*wasmbindingstypes.MsgCloseResponse, error) {
+func PerformMsgClose(f *marginkeeper.Keeper, ctx sdk.Context, contractAddr sdk.AccAddress, msgClose *types.MsgClose) (*types.MsgCloseResponse, error) {
 	if msgClose == nil {
 		return nil, wasmvmtypes.InvalidRequest{Err: "margin close null margin close"}
 	}
@@ -45,7 +45,7 @@ func PerformMsgClose(f *marginkeeper.Keeper, ctx sdk.Context, contractAddr sdk.A
 		return nil, errorsmod.Wrap(err, "margin close msg")
 	}
 
-	var resp = &wasmbindingstypes.MsgCloseResponse{
+	var resp = &types.MsgCloseResponse{
 		MetaData: msgClose.MetaData,
 	}
 	return resp, nil
